changed unsigned int type to unsigned long int

+++ libflow.c
@@ -6,7 +6,7 @@
 /*   By: ocarlos- <ocarlos-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/30 21:42:14 by ocarlos-          #+#    #+#             */
-/*   Updated: 2020/07/29 20:35:16 by ocarlos-         ###   ########.fr       */
+/*   Updated: 2020/07/30 11:03:26 by ocarlos-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ int		ft_vprintf(t_prtform *form, va_list lst, int count)
 {
 	char			*strarg;
 	int				argint;
-	unsigned int	argllint;
+	unsigned long int	argllint;   // changed here
 
 	argllint = 0;
 	if (form->typ == '%')
@@ -28,7 +28,7 @@ int		ft_vprintf(t_prtform *form, va_list lst, int count)
 	}
 	if (ft_testllint(form->typ))
 	{
-		argllint = va_arg(lst, unsigned int);
+		argllint = va_arg(lst, unsigned long int);   // changed here
 		form->siz = ft_redllint(argllint, form);
 	}
 	if (form->typ == 's')
@@ -53,7 +53,7 @@ int		ft_redint(int n, t_prtform *form)
 	return (form->siz);
 }
 
-int		ft_redllint(unsigned int n, t_prtform *form)
+int		ft_redllint(unsigned long int n, t_prtform *form)  // changed here
 {
 	if (form->typ == 'u')
 		form->siz = ft_puin(n, form);
